ecommerce_analytics/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-admin==2.0.2
django-excel-response2==3.0.6
django-restframework==0.0.1
django-six==1.0.5
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
openpyxl==3.1.5
packaging==24.1
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
screen==1.0.1
six==1.16.0
sqlparse==0.5.1
TimeConvert==3.0.13
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
xlwt==1.3.0
